# perf(req): improve `c.req.query` performance