--- a/deno_dist/utils/url.ts
+++ b/deno_dist/utils/url.ts
@@ -50,6 +50,16 @@ export const getPathFromURL = (url: string, strict: boolean = true): string => {
   return result
 }
 
+export const getQueryStringFromURL = (url: string): string => {
+  const queryIndex = url.indexOf('?')
+  const fragmentIndex = url.indexOf('#')
+  const result =
+    queryIndex !== -1
+      ? url.substring(queryIndex, fragmentIndex !== -1 ? fragmentIndex : undefined)
+      : ''
+  return result
+}
+
 export const isAbsoluteURL = (url: string): boolean => {
   const match = url.match(URL_REGEXP)
   if (match) {