--- a/src/utils/url.test.ts
+++ b/src/utils/url.test.ts
@@ -1,4 +1,11 @@
-import { splitPath, getPattern, getPathFromURL, isAbsoluteURL, mergePath } from './url'
+import {
+  splitPath,
+  getPattern,
+  getPathFromURL,
+  isAbsoluteURL,
+  mergePath,
+  getQueryStringFromURL,
+} from './url'
 
 describe('url', () => {
   it('splitPath', () => {
@@ -58,6 +65,22 @@ describe('url', () => {
       expect(path).toBe('/hello')
       path = getPathFromURL('https://example.com/hello/hey/', false)
       expect(path).toBe('/hello/hey')
+    })
+  })
+
+  describe('getQueryStringFromURL', () => {
+    it('should return strings of query params', () => {
+      let queryString = getQueryStringFromURL('https://example.com/?foo=bar')
+      expect(queryString).toBe('?foo=bar')
+      queryString = getQueryStringFromURL('https://example.com/?foo=bar&foo2=bar2')
+      expect(queryString).toBe('?foo=bar&foo2=bar2')
+      queryString = getQueryStringFromURL('https://example.com/')
+      expect(queryString).toBe('')
+      queryString = getQueryStringFromURL('https://example.com/?foo=bar#fragment')
+      expect(queryString).toBe('?foo=bar')
+      // This specification allows that the string includes two `?` or more
+      queryString = getQueryStringFromURL('https://example.com/?foo=bar?foo2=bar2')
+      expect(queryString).toBe('?foo=bar?foo2=bar2')
     })
   })