I've improved `c.req.query` performance. Currently, `c.req.query` uses `new URL` for parsing query parameters strings. It is inefficient. In this PR, `c.req.query` uses just `URLSearchParams`, it becomes to be faster.

## Benchmarks

Scripts:

```ts
import { run, bench } from 'mitata'
import { getQueryStringFromURL } from '../../src/utils/url'

const urlString = 'http://localhost?foo=bar&foo2=bar2&foo3=bar3'

{
  bench('URL', () => {
    const url = new URL(urlString)
    url.searchParams.get('foo')
  })
}

{
  bench('getQueryStringFromURL + URLSearchParams', () => {
    const queryStirng = getQueryStringFromURL(urlString)
    const searchParams = new URLSearchParams(queryStirng)
    searchParams.get('foo')
  })
}

await run()
```

Results:

```plain
MacBook@yusuke $ bun run ./url-params.ts
cpu: Apple M1 Pro
runtime: bun 0.1.11 (arm64-darwin)

benchmark                                    time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------------------------- -----------------------------
URL                                        1.18 µs/iter   (933.55 ns … 3.21 µs)   1.14 µs   3.21 µs   3.21 µs
getQueryStringFromURL + URLSearchParams  802.39 ns/iter   (696.65 ns … 2.35 µs) 795.49 ns   2.35 µs   2.35 µs

---

MacBook@yusuke $ deno run -A ./deno.ts
cpu: Apple M1 Pro
runtime: deno 1.25.1 (aarch64-apple-darwin)

benchmark                                    time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------------------------- -----------------------------
URL                                        2.62 µs/iter     (2.53 µs … 2.67 µs)   2.64 µs   2.67 µs   2.67 µs
getQueryStringFromURL + URLSearchParams    1.33 µs/iter      (1.3 µs … 1.34 µs)   1.33 µs   1.34 µs   1.34 µs
```

In the real world:

```ts
import { Hono } from '../../src/hono'

const app = new Hono()

app.get('/', (c) => {
  const q = c.req.query('q')
  return c.text(q)
})

export default app
```

Results:

Bun

```plain
---
v2.1.3

MacBook@yusuke $ bombardier -c 200 -d 10s 'http://localhost:3000/?q=search-query&foo=bar&foo2=bar2'
Bombarding http://localhost:3000/?q=search-query&foo=bar&foo2=bar2 for 10s using 200 connection(s)
[=======================================================================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec     93868.04   25880.52  119351.49
  Latency        2.13ms     1.26ms    26.70ms
  HTTP codes:
    1xx - 0, 2xx - 938298, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    16.73MB/s

After

MacBook@yusuke $ bombardier -c 200 -d 10s 'http://localhost:3000/?q=search-query&foo=bar&foo2=bar2'
Bombarding http://localhost:3000/?q=search-query&foo=bar&foo2=bar2 for 10s using 200 connection(s)
[=======================================================================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec    123577.17   30821.98  148889.77
  Latency        1.62ms     1.45ms    44.68ms
  HTTP codes:
    1xx - 0, 2xx - 1235784, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    22.04MB/s
```

Deno

```plain
---
v2.1.3

MacBook@yusuke $ bombardier -c 200 -d 10s 'http://localhost:9000/?q=search-query&foo=bar&foo2=bar2'
Bombarding http://localhost:9000/?q=search-query&foo=bar&foo2=bar2 for 10s using 200 connection(s)
[=======================================================================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec    105378.02    8459.58  117360.28
  Latency        1.90ms   229.71us    22.57ms
  HTTP codes:
    1xx - 0, 2xx - 1053922, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    22.51MB/s
  
---
  
After

MacBook@yusuke $ bombardier -c 200 -d 10s 'http://localhost:9000/?q=search-query&foo=bar&foo2=bar2'
Bombarding http://localhost:9000/?q=search-query&foo=bar&foo2=bar2 for 10s using 200 connection(s)
[=======================================================================================================================================================] 10s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec    136321.87   29600.10  168217.67
  Latency        1.47ms   165.65us    21.77ms
  HTTP codes:
    1xx - 0, 2xx - 1363227, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    29.12MB/s
```