@yusukebe 2022-09-05
@usualoma Thank you!

I've made it allows the fragments as query strings. The code became more simple.